+++ flutter_file_manager_platform_interface/lib/src/platform_interface.dart
@@ -1,17 +1,33 @@
+import 'dart:typed_data';
+
 import 'mime_type.dart';
 
 abstract class FileManagerPlatform {
   static FileManagerPlatform instance = FileManagerPlatformException();
 
-  /// Write [data] inside [fileName] to [path] and return the path to the file.
+  /// Write [data] inside [fileName] and return the path to the file.
   ///
   /// - `fileName`: The name of the file (e.g: `my_file.json`).
   /// - `data`: The data to write inside the file.
   /// - `type`: The type of the file, used for the web implementation. If not
   ///  provided, the type will be guessed by parsing the `fileName` extension.
-  Future<String> writeFile({
+  Future<String> writeFileAsString({
     required String fileName,
     required String data,
+    MimeType? type,
+  }) {
+    throw UnimplementedError('writeFile() has not been implemented.');
+  }
+
+  /// Write [bytes] inside [fileName] and return the path to the file.
+  ///
+  /// - `fileName`: The name of the file (e.g: `my_file.json`).
+  /// - `bytes`: The data to write inside the file.
+  /// - `type`: The type of the file, used for the web implementation. If not
+  ///  provided, the type will be guessed by parsing the `fileName` extension.
+  Future<String> writeFile({
+    required String fileName,
+    required Uint8List bytes,
     MimeType? type,
   }) {
     throw UnimplementedError('writeFile() has not been implemented.');