+++ flutter_file_manager_platform_interface/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1
+
+* Changed method `writeFile` to `writeFileAsString`
+* Created new method `writeFile` which takes a `Uint8List`
+* Fixed documentation
+
 ## 1.0.0
 
 * Initial release.